refactor(canvas): namespace render chunk coverage
