+++ packages/core/src/canvas/renderer/chunks/index.ts
